Prediction Framework Update
This is largely just a push of the attempts to design a solid foundation
for further ML predictions, however it has one testing method for a ML
Prediction already, and it has produced interesting results on the
training data it has been given. Admittedly, the high score is likely
just training data bias, but its enough to warrant further testing.